Update algDef.py
algDef now can be used in a real game of Connect4.  Some minor adjustments need to be made in order for it to not skip a turn.  However, it now "copies" the main board into a temporary board which will then calculate edges and pick the biggest value based on its rules